Move to separate resolver
