+++ components/console-backend-service/internal/domain/k8s/deployment_converter.go
@@ -1,10 +1,8 @@
 package k8s
 
 import (
-	"github.com/blang/semver"
 	"github.com/kyma-project/kyma/components/console-backend-service/internal/gqlschema"
 	api "k8s.io/api/apps/v1"
-	"strings"
 )
 
 type deploymentConverter struct{}
@@ -75,29 +73,4 @@ func (c *deploymentConverter) toGQLContainers(in api.Deployment) []gqlschema.Con
 	}
 
 	return containers
-}
-
-func (c *deploymentConverter) ToKymaVersion(deploymentImage string) string {
-	deploymentImageSeparated := strings.FieldsFunc(deploymentImage, split)
-
-	source := deploymentImageSeparated[0]
-	if source != "eu.gcr.io" {
-		return deploymentImage
-	}
-
-	version := deploymentImageSeparated[len(deploymentImageSeparated)-1]
-	_, err := semver.Parse(version)
-	if err != nil {
-		branch := "master"
-		if strings.HasPrefix(version,"PR-") {
-			branch = "pull request"
-		}
-		return strings.Join([]string{branch, version}, " ")
-	}
-
-	return version
-}
-
-func split(r rune) bool {
-	return r == '/' || r == ':'
 }
