+++ components/console-backend-service/internal/domain/k8s/k8s.go
@@ -30,6 +30,7 @@ type Resolver struct {
 	*replicaSetResolver
 	*configMapResolver
 	*selfSubjectRulesResolver
+	*kymaVersionResolver
 	informerFactory informers.SharedInformerFactory
 }
 
@@ -56,6 +57,10 @@ func New(restConfig *rest.Config, informerResyncPeriod time.Duration, applicatio
 	if err != nil {
 		return nil, errors.Wrap(err, "while creating deployment service")
 	}
+	kymaVersionService, err := newKymaVersionService(informerFactory.Apps().V1().Deployments().Informer())
+	if err != nil {
+		return nil, errors.Wrap(err, "while creating kymaVersion service")
+	}
 	limitRangeService := newLimitRangeService(informerFactory.Core().V1().LimitRanges().Informer(), clientset.CoreV1())
 
 	resourceService := newResourceService(clientset.Discovery())
@@ -81,6 +86,7 @@ func New(restConfig *rest.Config, informerResyncPeriod time.Duration, applicatio
 		resourceQuotaStatusResolver: newResourceQuotaStatusResolver(resourceQuotaStatusService),
 		configMapResolver:           newConfigMapResolver(configMapService),
 		selfSubjectRulesResolver:    newSelfSubjectRulesResolver(selfSubjectRulesService),
+		kymaVersionResolver:         newKymaVersionResolver(kymaVersionService),
 		informerFactory:             informerFactory,
 	}, nil
 }
